Updated python API
The update of the python API is complete for regions. Apparently, everything is working. The only difference between these results and those of ca2b90 commit are when the region has an angle different from zero.

Changes in pattern.c are to make it compliant with SuiteSparse. 

The make related files are forced to use the system gcc for compilation and linking.